v0.3.5 History uniqueness
Since everything in python is a reference if you calculate or display LaTeX or plot an item with the contextmenu which is already in the corresponding history a reference would be created and added to the history, thus the same item would be shown twice. since this is madness and clogs up the history.
+ An item will only be shown once per history and if it is called in the tab again its entry in the history will move to the bottom of the history.