v0.9.0: The Multidimensional Update!!!
NEW TAB: Tab 4 allows the input of Vectors and Maticies to use them in Equations. This can also be used to solve systems of linear equations.
With this, AMaDiA is capable of everything that one can encounter in school-math and even a bit more.

The Tab is still WIP and should be used with caution!

+ Added Tab 4
+ Restructured the AMaS Class to support everything