Extract Tab 1
Tab 1 is now in its own file.
The goal is to get rid of the UI file and to split all of the UI building and functionality of the tabs into their own files to make it easier to organise everything and to make changes to a single tab without needing to rebuild the ui file every time.