Extract the other Tabs
All tabs are now in their own files with UI and methods in one place so that they can be changed easier.
(The widgets and Methods might need a little bit of further renaming.)
The next task is to do this this to the tabs i.e. the tabs need to be further separated into widgets like a settings widget for the 2D plot as the code is still a mess (but at least a mess that can be worked with now).