Factorial Plots
I made some modifications to how the output of lambdify is used to generate the y-values to be more robust when handling functions.
I also added handling of ValueErrors by trying to plot only for positive x values to make the usage more intuitive when plotting functions that are only defined for positive values.