LaTeX but with Unicode
The solver can now handle LaTeX that contains common unicode symbols by replacing those with the corresponding LaTeX commands.
The warning "Could not convert input to LaTeX" now no longer pops up when the input already is LaTeX.
Adds check for number of braces and brackets (in addition to the already existing check for parentheses) when deciding whether an equation is queried to handle LaTeX sums and similar constructs correctly.
The Multiline LaTeX input now handles each line separately and checks if a given line is already valid LaTeX if conversion to LaTeX fails. This allows the use of the LaTeX tab as a simple LaTeX display for LaTeX code in addition to its main function as a LaTeX converter.