Experimental Engineering Matrix
I have begun work on an experimental engineering matrix class that behaves much more like a numpy matrix instead of the sympy matrix in terms of elementwise and scalar operations. This is currently very much work in progress and does not really work yet, except that it successfully performs elementwise multiplication with "*" and matmul with "@" as an operator smybol.
The use of this new matrix is controlled via a developer option in the options window.